refactor(ipc): split projectCrud.ts into focused sub-domain handlers
- Extract into electron/ipc/handlers/projectCrud/: crud, switch, settings, stats, gitInit, gitClone
- Keep projectCrud.ts as a backward-compat shim that re-exports from projectCrud/index.js (mirrors terminal.ts pattern)
- Collapse duplicated multi-view/PTY/worktree-port orchestration shared by switch + reopen into private persistOutgoingProjectState and activateProjectView helpers in switch.ts
- Preserve IPC channel names, getProjectStatsService export, and the cloneAbortController / projectStatsServiceInstance module-level state inside their owning sub-files
